Some fixes that will be part of oXygen 18.1
Headers of simple tables and properties tables are repeated on following
pages
The @navtitle attribute from the booklists/toc element is used to
generate the title of the TOC section.
The width set on the table colspec element is used in the PDF output.
Does not break a simple table row over two pages.
Some fixes for the hierarchic indexterms
Support for the index-see and index-see-also
Image references to remote resources (web) are now supported.